Add partials for header/footer, migrations for users, orders, cart
This code breaks the HTML for the header and footer out into
their own partials to be more easily managed.

It also adds the functionality for the cart. The cart is a class
that doesn't inherit from ActiveRecord::Base. When the application
starts up, in the application_controller, it instantiates a new cart
that is essentially an empty hash that can store the items the user
adds to the cart.

Another part that was added, which goes along with the cart functionality
is the session.

This code also adds the users along with validations on that model
as well as the orders and the order_products join table.